added loader for saving state
