+++ health/micro-ui/web/packages/modules/campaign-manager/src/components/ComponentToRender.js
@@ -57,7 +57,7 @@ const ComponentToRender = ({ field, t: customT, selectedField, isSelected }) => 
         description={shouldCustomTranslate ? field?.helpText : customT(field?.helpText) || ""}
         error={shouldCustomTranslate ? field?.errorMessage : customT(field?.errorMessage) || null}
         infoMessage={shouldCustomTranslate ? field?.tooltip : customT(field?.tooltip) || null}
-        label={shouldCustomTranslate ? field?.label : customT(field?.label) || ""}
+        label={field?.showLabel === false ? "" : shouldCustomTranslate ? field?.label : customT(field?.label) || ""}
         onChange={function noRefCheck() {}}
         placeholder={shouldCustomTranslate ? field?.innerLabel : customT(field?.innerLabel) || ""}
         populators={{
